chore(agent@rc.6): move chain tester to fixtures
- testChainSupport -> testConnectionWithBackend
