+++ agent/agent.ts
@@ -26,4 +26,3 @@ export type { CodeId, CodeHash, } from './deploy'
 export * as Token from './token'
 export type { Uint128, Uint256, Decimal128, Decimal256 } from './token'
 export * as Stub from './stub'
-export * as Tester from './tester'
