+++ agent/package.json
@@ -1,6 +1,6 @@
 {
   "name":    "@fadroma/agent",
-  "version": "2.0.0-rc.5",
+  "version": "2.0.0-rc.6",
   "type":    "module",
   "main":    "agent.ts",
   "files": [ "*.ts", "!*.test.*" ],
