+++ fixtures/package.json
@@ -1,5 +1,4 @@
 {
   "name": "@fadroma/fixtures",
-  "private": true,
   "main": "fixtures.ts"
 }
